+++ pydreamplet/types.py
@@ -1,0 +1,2 @@
+type Real = int | float
+type NumericPair = tuple[Real, Real] | list[Real]